Refactor mixins
